Added product upload controller
